Update patch set 9
Patch Set 9:

On the examples I've looked at so far, when fedora-07 passes, it's only because the vhost target core dumped - and the core dump was directly related to the issues you're fixing here.

Looking at the results from this latest revision, fedora-07 passed, but no core dump from the vhost app.

Patch-set: 9