Update patch set 3
Patch Set 3:

Thanks for the comments.

For the suggestion to have a single queue for the nomem condition and IO throttling (QoS) condition, it looks like not a good idea here. These are two different cases and there is no necessary to consolidate into one.
"submit_bdev_io(bdev_io):
    if (nomem_io queue is not empty || io is throttled)
        insert  bdev_io to nomem_io queue
    else
        submit(bdev_io)"

On the other hand, the nomem IO queue is for each bdev channel and QoS IO queue is only for the master channel. The QoS functionality can be enabled and disabled at the runtime dynamically. So have a specific IO queue for a specific master channel looks like more clear and reasonable.

Again, thanks for the comments.

Patch-set: 3
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0