Update patch set 5
Patch Set 5: Code-Review-1

(4 comments)

I fail to see how this BDAL guide can help me at `supplementing the source code to provide an overall understanding [...]`.

Please try to read and understand this guide from a user perspective.
In my opinion it badly needs an example with step-by-step, in-depth explanation of what's going on. It doesn't need a simplified copy of spdk_bdev_fn_table struct definition. The spdk_bdev struct should be mentioned a bit earlier than in the very last section of the guide.

You might want to rebase this patch on Ben's guide for io_channels and add a bunch of @refs. io_channels are very crucial here.

Patch-set: 5
Label: Code-Review=-1