Update patch set 3
Patch Set 3:

WeWe, is your intention to propose this as part of the SPDK repo or do you intent to maintain this in your own repo and are simply asking for review comments? It's not quite clear based on the feedback given previously and you recent email.

Patch-set: 3
Label: Code-Review=0