Update patch set 1
Patch Set 1: Code-Review-1

Jim and I discussed this at length. I think there is probably a better long term solution that would require a ton of refactoring, but for 18.01 I think this is the right thing to do. Some of the names can be clarified though.

1) Instead of shared channel, how about "module channel"? It's associated with the module.
2) Instead of shared_ch->owner how about shared_ch->base_ch or shared_ch->module_ch?

We'll definitely need to do some stress testing to make sure this is working.

Patch-set: 1
Label: Code-Review=-1