Update patch set 1
Patch Set 1: Code-Review+1

I think that this is a good fix for the failures you were facing on fedora-08. Especially since it is clear from the fact that you only experiences the failure on the physical machine and not the virtual machine that the system is just getting into a weird state and there is nothing wrong with the test itself.

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1