Update patch set 20
Patch Set 20:

Thought I posted this already, maybe I commented on another review? Anyway, 2 questions (1) is the process to push a separate patch to test nightly or modify the one in question twice, once to enable and then again to disable? I'd think the second one and (2) looking at the nightly run of this we're adding 10 min to the test on 2 machines. Given this is nightly so not as critical but the comment has been made that night in the US is not night everywhere else :) Unless I'm reading the chiming charts wrong this patch is a pretty significant increase and maybe we want to cut down on the test run times or # of cases or something?

Patch-set: 20
Label: Code-Review=0