Update patch set 20
Patch Set 20:

395015 looks like a duplicate of this patch, with RUN_NIGHTLY enabled, and it passed.  So it looks like it did pass the test pool.

But it would be better to just modify this patch with the RUN_NIGHTLY changes, including a change to the commit message like

NIGHTLY: autotest: vhost_fio_test (porting from nightly to autotest)

Then the results for the NIGHTLY run are part of this same review.

I'm adding Karol Latecki also to review this patch since it is related to vhost.  Karol - please include anyone else you would like to review it.

Please note I have not reviewed the patch myself yet - will wait for others to +1 first.

Thanks,

-Jim

Patch-set: 20