spdk_phys_regions: Physical regions for vtophys
In order to permit vtophys translations for physical memory like NVMe CMBs
and PMRs we create a new TAILQ structure called g_phys_regions and
populate it via calls to the spdk_mem_register function. vtophys
lookups then use a new function (spdk_vtophys_phys()) when other
methods fail and if we find the vaddr inside an phys mapping we return
the physicial address based on the offset within the region.

Change-Id: I4f055bd9e2630f258e929ae7762da913550fc47d
Signed-off-by: Stephen Bates <sbates@raithlin.com>