Update patch set 2
Patch Set 2:

Another option which would help in some cases (i.e. SIGSEGV, SIGBUS) but not all (i.e. SIGKILL) is to install a SEGV handler and unlink the file there.  See http://www.alexonlinux.com/how-to-handle-sigsegv-but-also-generate-core-dump for some details on this - including making sure you still call the default handler so that a core dump gets generated.

Then at least for core dumps, the socket file should get unlinked.  kill -9 cannot be caught - so socket file would still need to be deleted manually in this case.

Shuhei - would this be acceptable and useful to you?

Patch-set: 2