Update patch set 40
Patch Set 40:

We also still need changes to alloc a cluster, but not put it into the cluster map, until after the copy is complete.  Otherwise we could have two separate channels racing on the same cluster.

That should be done in a separate patch though - let's keep this one with the current strategy.

Patch-set: 40
Label: Code-Review=0