bdev/qos: cleanup the IOs on QoS thread when removing IO channel
The I/O channel can be obtained and released dynamically on
demand, in the case of QoS has been enabled. The I/Os could
be queued on the QoS thread due to rate limiting. Considering
the thread safe operation, the thread for the I/O channel and
QoS thread is mostly different. The asynchronous method of
message passing through threads is used to drain the I/Os when
"spdk_put_io_channel()" is executed to cleanup the I/O channel.

In this way:
1. Add the async support for spdk_put_io_channel() API for this
asynchronous I/Os draining on QoS thread
2. Update the spdk_io_channel_destroy_cb() to return a value to
indicate the async destroy operation
3. Related changes for other places on the destroy_cb.
4. UT code added to release I/O channel at runtime

Change-Id: I056f211670be7e3d2f3b89c9dd46c94f46d1cb93
Signed-off-by: GangCao <gang.cao@intel.com>