bdev: Vbdev Passthru Example
*** WIP *** posting so I can an reference the code when asking a
few questions on IRC.

Example showing the basics of a vbdev via a module that implements
the main functions, and a few extras, needed for a vbdev. The tests
also demonstrate how to configure an application to use a vbdev and
an underlying bdev.

This will also be the basis for the DPDK-FW based crypto module.

Change-Id: I67bf542143eb6532465cdb6af63c0f2ab7084c0e
Signed-off-by: paul luse <paul.e.luse@intel.com>