Makefile: try harder to use correct Python interpreter
In a Python 3-only system, there might not be an executable in the
PATH to satisfy "python". In configure we attempt to find the "right"
python command for the system we happen to be running in. With this
patch, we let configure save this command to a temporary file.

Subsequently, when make is run, it will find the command in that file
and use it (or fall back to "python" if the file doesn't exist for
whatever reason).

Change-Id: Ib357413d51ea51492bf7a8ae8881ba502644dbcc
Signed-off-by: Nathan Cutler <ncutler@suse.com>