Update patch set 1
Patch Set 1:

Please disregard my earlier comment. The openSUSE python maintainers just told me that there are no plans to link /usr/bin/python to anything other than python2, so with the deprecation of Python 2 there is going to be a realistic possibility that /usr/bin/python might not be present.

So the question is not whether to deal with it here, but how.

Ideally it would be done in a way that could be applied everywhere we call out to the python interpreter, and not just here in the Makefile.

Patch-set: 1