Update patch set 4
Patch Set 4: Code-Review-1

This article is good - but it really explains how flash SSDs work and doesn't give any practical guidance on benchmarking.  The only benchmarking advice given is a link to a SNIA article.

So I think we should check this in but with a different title.  It would be good to have some practical flash SSD benchmarking guidelines too but we can do that in a separate doc commit.

Patch-set: 4
Label: Code-Review=-1