Update patch set 13
Patch Set 13:

Yes, I definitely return to this patch in this release. This was just a POC but if we agree on mailing list or trello details it will be easier for me to worn on this.

Patch-set: 13
Reviewer: Pawel Wodkowski <1011230@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0