Update patch set 18
Patch Set 18:

OK. I've finally pulled all of this together and updated the original patch series with a set of re-based changes. Thanks to Paul Luse for his help in making this happen.

There is more work to be done, but I wanted to put these patches upstream, even though they are incomplete - just in case someone else would like to help move this work forward.  Feel free to do that, if you are inclined.  I will not be offended.

TODO:

1. The histogram_ut tests fail because of the following error. However, I see this failure on my internal testbed even on the 
master branch.  So this might have nothing to do with this path.
(This is where I can use some help from the community.

+ valgrind --leak-check=full --error-exitcode=2 test/unit/include/spdk/histogram_data.h/histogram_ut
==40355== Memcheck, a memory error detector
==40355== Copyright (C) 2002-2013, and GNU GPL'd, by Julian Seward et al.
==40355== Using Valgrind-3.10.1 and LibVEX; rerun with -h for copyright info
==40355== Command: test/unit/include/spdk/histogram_data.h/histogram_ut
==40355== 
==40355== Conditional jump or move depends on uninitialised value(s)

2. We need to address some of Jim's comments.  I've cut them out of some private email and pasted them here.

I (jim) still have a few questions on this patch.  I’ll put these in GerritHub as well.
 
1)      I’m concerned about the accuracy of the data on a multi-threaded NVMf target app.  Currently I don’t see any synchronization on the global histogram data structures.  I think this needs histogram data structures per thread, and then iterators that aggregate the data when a user requests the updated histogram data.
2)      It would be great if this was more dynamic, so that it does not have to be compiled in with a separate flag.
3)      Have you measured the time it takes to iterate over a fairly well populated histogram and send its data over JSON-RPC?   Just want to make sure that using JSON-RPC for the raw histogram data is not going to result in too much overhead.  I’ve wondered if we should just be sending the raw data using some kind of encoding, and let the client app work out the ranges and counts instead.
 
Thanks,
 
-Jim

Patch-set: 18
Label: Code-Review=0