Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

Hi John and Girish,

One of the things we are actively looking at is providing histogram capabilities on a per-bdev basis.  If we added that capability, how much of this patch would still be needed?    I think at least one of the histograms could be eliminated (in addition to the two that are currently not used).

The comments from this previous patch really apply to this one so adding them here:

1)	I’m concerned about the accuracy of the data on a multi-threaded NVMf target app.  I didn’t see any synchronization on the global histogram data structures.  I think this needs histogram data structures per thread, and then iterators that aggregate the data when a user requests the updated histogram data.
2)	It would be great if this was more dynamic, so that it does not have to be compiled in with a separate flag.
3)	Have we measured the time it takes to iterate over a fairly well populated histogram and send its data over JSON-RPC?   Just want to make sure that using JSON-RPC for the raw histogram data is not going to result in too much overhead.  I’ve wondered if we should just be sending the raw data using some kind of encoding, and let the client app work out the ranges and counts instead.

Thanks,

-Jim

Patch-set: 9
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1