Update patch set 11
Patch Set 11:

(1 comment)

Hi Jim,

As the poller execution is not really handled by a background thread, it is manually called through the poll_thread() API, in a later patch from the QoS change, I've added a parameter to determine whether executing the poller. In this way, the IOs can always be queued due to QoS.

The overall effect is 1) the poller will always be executed in the poll_thread() API 2) the poller can be skipped in the poll_thread() when a flag is set.

I'll also look into Shuhei's comment. Thanks.

Patch-set: 11