Update patch set 3
Patch Set 3:

For this:

"For example, if the rate limiting is set to be an IO per poller_run and we can run the ut_poller twice, we can test if two IOs ran after poller ran."

In my current change:
1. Queue one IO on the list
2. Trigger the poller through set_thread_passed_ticks()
3. The IO get executed through poll_thread()
4. Repeat 1-3 for the second IO

Or the suggestion is that each call of poll_thread() to invoke the poller and only kick off one queued IO? How many IOs to send down is depending on the allowed QoS setting and not relate to the number of calls on poller.

Take a detailed look at the comment on https://review.gerrithub.io/#/c/393012/ patch which seems like what I have implemented here to check the passed ticks (current ticks) with the next ticks (period ticks)

Thanks.

Patch-set: 3