Update patch set 19
Patch Set 19:

Well that was fun :) . So when adding a 'match' feature to a test script it will take a few rounds on the CI systems to decide which output needs to genericized via token and which don't In this last case I thought keeping the metadata offset value in there was pretty safe but a rebase after my initial match output check-in proved me wrong. This value did change and is probably not something we need to get a trigger on when it does but was a great example of the power of using this match thing. It can catch any level of output changes that we want!

This is ready for review now.

Patch-set: 19
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0