rte_vhost/fd_man: remove fdentry->busy flag
Now that we're sure there is a running
thread polling polling each fdentry, we
can leave the job of deleting fdentries
up to that thread. This leaves us with
one synchronization primitive less. It
also allows us to call fdset_del()
function inside the fd poll callback and
simplifies the API in general.

Change-Id: I676d9c22df6b07160d60efa3cf129a6780dc33f5
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>