Update patch set 1
Patch Set 1:

How important is this patch?  I guess I'd prefer to avoid extra changes to our copy of DPDK rte_vhost until we bottom out on our strategy.  We have enough differences already.

FYI - QEMU and DPDK are working on ability to enable a process in a VM to act as the vhost target for another VM.  We will want the changes they are adding to DPDK rte_vhost to enable this.

Patch-set: 1
Label: Code-Review=0