Update patch set 1
Patch Set 1:

An spdk_sock can only be of one type - for example, spdk_posix_sock or (in the future) spdk_vpp_sock.

But for spdk_sock_group, it may contain spdk_socks of different types.  This is the reason for spdk_sock_group_impl.  With POSIX + VPP, we create two spdk_sock_group_impls - one for POSIX, one for VPP - and attach both to the spdk_sock_group.  Then when spdk_sock is added, we make sure we pass it to the correct spdk_sock_group_impl.

Hopefully this explains it better.  Obviously I will be writing better documentation for all of this.  :)

Patch-set: 1
Label: Code-Review=0