Update patch set 5
Patch Set 5: Code-Review+2

This looks good to me.

However, since this breaks compatibility with existing RPC users, we should definitely add a note to CHANGELOG.md so that users are notified.

We are planning to make a significant number of RPC method changes to enable the JSON configuration format in the near future, so I think this can be a separate top-level notice in the changelog that just mentions that backwards compatibility may be broken for any/all RPC methods in this release, with the expectation that all RPC changes after this release will strive to be backwards compatible again.

Patch-set: 5
Reviewer: Daniel Verkamp <1010525@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2